fix: block-use mixin, shared buffers, disconnect cleanup, ID reuse
Critical fixes from code review:
- BlockUseSoundMixin: target ServerPlayerInteractionManager.interactBlock()
  instead of Block.onUse() which subclasses override without super call
- DetectionTracker: try/finally around settingDetectedTarget flag
- SearchLastKnownPositionGoal: create PacketByteBuf per player to avoid
  buffer reuse after send
- MobSetTargetMixin: send spotted packet to all tracking players, not just
  the target; new buffer per player

Correctness fixes:
- Client disconnect cleanup: clear SpottedTracker, DetectionBarRenderer,
  ClientModConfig, StealthCalculator on disconnect
- SpottedTracker: searchingMobs now uses TTL (30s) to prune stale entries
  from despawned mobs
- DetectionTracker + SoundFatigue: use mob UUID instead of entity ID to
  prevent data leaking to newly spawned mobs via ID reuse
- StealthCalculator: use HostileEntity filter instead of hearing filter
  for proximity/FOV checks (mobs that can see but not hear were excluded)
- BlockHitSoundMixin: remove dead code in action filter
- DetectionBarRenderer: clientTick as long to prevent overflow

+++ src/main/java/com/dysmn/doyouseemenow/client/DetectionBarRenderer.java
@@ -28,8 +28,8 @@ public final class DetectionBarRenderer {
     private static final Map<Integer, Float> displayProgress = new HashMap<>();
 
     /** Tick counter for stale entry cleanup. */
-    private static final Map<Integer, Integer> lastUpdateTick = new HashMap<>();
-    private static int clientTick = 0;
+    private static final Map<Integer, Long> lastUpdateTick = new HashMap<>();
+    private static long clientTick = 0;
 
     private static final float LERP_SPEED = 0.25f;
     private static final int STALE_TIMEOUT = 40; // 2 seconds
